# docs/api/openapi.yaml
openapi: 3.1.0
info:
  title: NeonVibe API
  description: "Servidor de música personal self-hosted. Los endpoints bajo /api/v1\
    \ requieren un JWT de acceso (header Authorization: Bearer <token>), salvo /auth/**\
    \ y /public/**."
  version: v1
servers:
- url: https://neonvibe.fepdev.app
  description: Producción
security:
- bearer: []
paths:
...
  /api/v1/playlists/{id}:
    get:
      tags:
      - playlist-controller
      operationId: getById
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int64
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/PlaylistDetailResponse"
    put:
      tags:
      - playlist-controller
      operationId: update_2
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int64
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PlaylistRequest"
        required: true
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/PlaylistResponse"
    delete:
      tags:
      - playlist-controller
      operationId: delete
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int64
      responses:
        "200":
          description: OK
...
  /api/v1/playlists:
    get:
      tags:
      - playlist-controller
      operationId: list
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/PlaylistResponse"
    post:
      tags:
      - playlist-controller
      operationId: create
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PlaylistRequest"
        required: true
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/PlaylistResponse"
  /api/v1/playlists/{id}/tracks:
    post:
      tags:
      - playlist-controller
      operationId: addTrack
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int64
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PlaylistTrackRequest"
        required: true
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/PlaylistResponse"
  /api/v1/playlists/{id}/reorder:
    post:
      tags:
      - playlist-controller
      operationId: reorder
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int64
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ReorderRequest"
        required: true
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/PlaylistResponse"
...
  /api/v1/history:
    get:
      tags:
      - history-controller
      operationId: list_1
      parameters:
      - name: pageable
        in: query
        required: true
        schema:
          $ref: "#/components/schemas/Pageable"
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/PagePlayHistoryResponse"
    post:
      tags:
      - history-controller
      operationId: record
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PlayHistoryRequest"
        required: true
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/PlayHistoryResponse"
...
  /api/v1/tracks:
    get:
      tags:
      - track-controller
      operationId: list_3
      parameters:
      - name: q
        in: query
        required: false
        schema:
          type: string
      - name: artist
        in: query
        required: false
        schema:
          type: string
      - name: album
        in: query
        required: false
        schema:
          type: string
      - name: genre
        in: query
        required: false
        schema:
          type: string
      - name: year
        in: query
        required: false
        schema:
          type: integer
          format: int32
      - name: pageable
        in: query
        required: true
        schema:
          $ref: "#/components/schemas/Pageable"
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/PageTrackResponse"
  /api/v1/tracks/{id}:
    get:
      tags:
      - track-controller
      operationId: getById_1
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int64
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/TrackResponse"
  /api/v1/tracks/{id}/stream:
    get:
      tags:
      - stream-controller
      operationId: stream
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int64
      - name: Range
        in: header
        required: false
        schema:
          type: string
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                type: string
                format: binary
...
  /api/v1/public/playlists/{id}:
    get:
      tags:
      - public-controller
      operationId: get_2
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int64
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/PublicPlaylistResponse"
  /api/v1/lastfm/callback:
    get:
      tags:
      - last-fm-controller
      operationId: callback
      parameters:
      - name: token
        in: query
        required: true
        schema:
          type: string
      responses:
        "200":
          description: OK
...
  /api/v1/playlists/{id}/tracks/{trackId}:
    delete:
      tags:
      - playlist-controller
      operationId: removeTrack
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int64
      - name: trackId
        in: path
        required: true
        schema:
          type: integer
          format: int64
      responses:
        "200":
          description: OK
...
components:
  schemas:
...
    PlaylistRequest:
      type: object
      properties:
        name:
          type: string
          maxLength: 200
          minLength: 0
        description:
          type: string
          maxLength: 2000
          minLength: 0
        isPublic:
          type: boolean
      required:
      - name
    PlaylistResponse:
      type: object
      properties:
        id:
          type: integer
          format: int64
        name:
          type: string
        description:
          type: string
        isPublic:
          type: boolean
        coverArtPath:
          type: string
        ownerId:
          type: string
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
        tracks:
          type: array
          items:
            $ref: "#/components/schemas/PlaylistTrackResponse"
    PlaylistTrackResponse:
      type: object
      properties:
        id:
          type: integer
          format: int64
        trackId:
          type: integer
          format: int64
        position:
          type: integer
          format: int32
    PlaylistTrackRequest:
      type: object
      properties:
        trackId:
          type: integer
          format: int64
      required:
      - trackId
    ReorderRequest:
      type: object
      properties:
        trackIds:
          type: array
          items:
            type: integer
            format: int64
      required:
      - trackIds
    PlayHistoryRequest:
      type: object
      properties:
        trackId:
          type: integer
          format: int64
        completed:
          type: boolean
        durationListenedSeconds:
          type: integer
          format: int32
      required:
      - trackId
    PlayHistoryResponse:
      type: object
      properties:
        id:
          type: integer
          format: int64
        trackId:
          type: integer
          format: int64
        playedAt:
          type: string
          format: date-time
        completed:
          type: boolean
        durationListenedSeconds:
          type: integer
          format: int32
...
    Pageable:
      type: object
      properties:
        page:
          type: integer
          format: int32
          minimum: 0
        size:
          type: integer
          format: int32
          minimum: 1
        sort:
          type: array
          items:
            type: string
    PageTrackResponse:
      type: object
      properties:
        totalElements:
          type: integer
          format: int64
        totalPages:
          type: integer
          format: int32
        size:
          type: integer
          format: int32
        content:
          type: array
          items:
            $ref: "#/components/schemas/TrackResponse"
        number:
          type: integer
          format: int32
        sort:
          $ref: "#/components/schemas/SortObject"
        pageable:
          $ref: "#/components/schemas/PageableObject"
        numberOfElements:
          type: integer
          format: int32
        first:
          type: boolean
        last:
          type: boolean
        empty:
          type: boolean
    PageableObject:
      type: object
      properties:
        offset:
          type: integer
          format: int64
        sort:
          $ref: "#/components/schemas/SortObject"
        paged:
          type: boolean
        pageNumber:
          type: integer
          format: int32
        pageSize:
          type: integer
          format: int32
        unpaged:
          type: boolean
    SortObject:
      type: object
      properties:
        empty:
          type: boolean
        sorted:
          type: boolean
        unsorted:
          type: boolean
    TrackResponse:
      type: object
      properties:
        id:
          type: integer
          format: int64
        filePath:
          type: string
        title:
          type: string
        artist:
          type: string
        album:
          type: string
        albumArtist:
          type: string
        year:
          type: integer
          format: int32
        genre:
          type: string
        trackNumber:
          type: integer
          format: int32
        discNumber:
          type: integer
          format: int32
        durationSeconds:
          type: integer
          format: int32
        bitrate:
          type: integer
          format: int32
        format:
          type: string
        mimeType:
          type: string
        hasLyrics:
          type: boolean
        coverArtPath:
          type: string
        isAvailable:
          type: boolean
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
...
    PublicPlaylistResponse:
      type: object
      properties:
        id:
          type: integer
          format: int64
        name:
          type: string
        description:
          type: string
        ownerId:
          type: string
        tracks:
          type: array
          items:
            $ref: "#/components/schemas/PublicTrackResponse"
    PublicTrackResponse:
      type: object
      properties:
        id:
          type: integer
          format: int64
        title:
          type: string
        artist:
          type: string
        album:
          type: string
        year:
          type: integer
          format: int32
        genre:
          type: string
        trackNumber:
          type: integer
          format: int32
        durationSeconds:
          type: integer
          format: int32
        hasLyrics:
          type: boolean
    PlaylistDetailResponse:
      type: object
      properties:
        id:
          type: integer
          format: int64
        name:
          type: string
        description:
          type: string
        isPublic:
          type: boolean
        coverArtPath:
          type: string
        ownerId:
          type: string
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
        tracks:
          type: array
          items:
            $ref: "#/components/schemas/TrackResponse"
    PagePlayHistoryResponse:
      type: object
      properties:
        totalElements:
          type: integer
          format: int64
        totalPages:
          type: integer
          format: int32
        size:
          type: integer
          format: int32
        content:
          type: array
          items:
            $ref: "#/components/schemas/PlayHistoryResponse"
        number:
          type: integer
          format: int32
        sort:
          $ref: "#/components/schemas/SortObject"
        pageable:
          $ref: "#/components/schemas/PageableObject"
        numberOfElements:
          type: integer
          format: int32
        first:
          type: boolean
        last:
          type: boolean
        empty:
          type: boolean
...
  securitySchemes:
    bearer:
      type: http
      name: bearer
      scheme: bearer
      bearerFormat: JWT